+++ src/components/profile/MyComments.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const MyComments = () => {
+  return (
+    <div>MyComments</div>
+  )
+}
+
+export default MyComments
