+++ src/components/profile/MyContents.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const MyContents = () => {
+  return (
+    <div>MyContents</div>
+  )
+}
+
+export default MyContents
